+++ build.sh
@@ -1,3 +1,0 @@
-dart run build_runner build --delete-conflicting-outputs
-dart format example/lib/pubspec.dart
-dart format lib/pubspec.dart