+++ lib/src/platform/platform.dart
@@ -1,1 +1,1 @@
-export 'platform_io.dart' if (dart.library.html) 'platform_web.dart';
+export 'platform_web.dart' if (dart.library.io) 'platform_io.dart';