+++ lib/src/platform/platform_io.dart
@@ -1,6 +1,9 @@
 import 'dart:io';
 
+import 'package:mcp_dart/mcp_dart.dart' as mcp;
+
 import '../agent/agent.dart';
+import '../agent/mcp_client.dart';
 
 /// Gets the value of an environment variable.
 ///
@@ -20,3 +23,36 @@ String getEnv(String key) {
   }
   return value;
 }
+
+/// Gets the transport for the MCP server.
+mcp.Transport getTransport({
+  required McpServerKind kind,
+  required String? command,
+  required List<String> args,
+  required Map<String, String> environment,
+  required String? workingDirectory,
+  required Uri? url,
+  Map<String, String>? headers,
+  Map<String, String>? requestInit,
+}) => switch (kind) {
+  McpServerKind.local => mcp.StdioClientTransport(
+    mcp.StdioServerParameters(
+      command: command!,
+      args: args,
+      environment: environment,
+      workingDirectory: workingDirectory,
+      // Fix mcp_dart bug: must use normal mode to access stdout/stdin
+      // pipes
+      stderrMode: ProcessStartMode.normal,
+    ),
+  ),
+  McpServerKind.remote => mcp.StreamableHttpClientTransport(
+    url!,
+    opts:
+        headers == null
+            ? null
+            : mcp.StreamableHttpClientTransportOptions(
+              requestInit: {'headers': headers},
+            ),
+  ),
+};